Add switch for enhancement mode
