+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -359,6 +359,14 @@
     <string name="install_known_issue_details">详细信息…</string>
     <string name="about_libraries_title">开放源代码许可</string>
     <string name="default_changes_log">&lt;b&gt;暂无更新日志&lt;/b&gt;&lt;br /&gt;美好的事物都值得等待</string>
-    <string name="xposed_description">启用 EdXposed 的增强功能:\n判断 Xposed 是否正常工作\n移除针对 EdXposed 的后台限制\n假装安装了 Xposed Installer</string>
+    <string name="xposed_description">启用 EdXposed 的增强功能:\n判断 Xposed 是否正常工作\n移除针对 EdXposed 的后台限制\n假装安装了 Xposed Installer\n隐藏 EdXposed Manager</string>
     <string name="status_enhancement">增强模式已激活</string>
+    <string name="settings_group_enhancement">增强模式</string>
+    <string name="settings_title_enhancement">增强模式状态</string>
+    <string name="settings_summary_enhancement">未激活\n你可以在「模块」中启用增强模块</string>
+    <string name="settings_summary_enhancement_enabled">已激活</string>
+    <string name="settings_title_pretend_xposed_installer">Xposed Installer 伪装</string>
+    <string name="settings_summary_pretend_xposed_installer">假装已安装 Xposed Installer 来使一些过时但有用的模块正常工作\n注：开启此功能可能会被某些软件检测到 EdXposed</string>
+    <string name="settings_title_hide_edxposed_manager">隐藏 EdXposed Manager</string>
+    <string name="settings_summary_hide_edxposed_manager">防止软件检测到 EdXposed Manager\n注：模块可能无法正常打开 Manager 界面</string>
 </resources>
