+++ app/src/main/res/xml/prefs.xml
@@ -237,4 +237,27 @@
 
     </PreferenceCategory>
 
+    <PreferenceCategory
+        android:key="group_enhancement"
+        android:title="@string/settings_group_enhancement">
+
+        <Preference
+            android:key="enhancement_status"
+            android:title="@string/settings_title_enhancement"
+            android:summary="@string/settings_summary_enhancement" />
+
+        <SwitchPreference
+            android:defaultValue="false"
+            android:key="pretend_xposed_installer"
+            android:summary="@string/settings_summary_pretend_xposed_installer"
+            android:title="@string/settings_title_pretend_xposed_installer" />
+
+        <SwitchPreference
+            android:defaultValue="false"
+            android:key="hide_edxposed_manager"
+            android:summary="@string/settings_summary_hide_edxposed_manager"
+            android:title="@string/settings_title_hide_edxposed_manager" />
+
+    </PreferenceCategory>
+
 </PreferenceScreen>
